+++ app/lib/encryption.server.js
@@ -1,0 +1,5 @@
+import Cryptr from "cryptr";
+
+const encryption = new Cryptr(process.env.ENCRYPTION_STRING);
+
+export default encryption;